Factor out a MetadataStore so the controller doesn't know about storage
